Add files via upload
Added the ability to restart the config without exiting Neovim. It has a lock file so that you can't reload thousands of times per second (which would devour the CPU and RAM, and crash the system... I found that out the hard way). Restarting is done with the user command "ReloadNvim"